backend/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.34.*
pydantic-settings==2.*
openai==1.*
httpx==0.28.*
boto3==1.*
faiss-cpu==1.*
numpy==2.*
python-multipart==0.0.*
websockets==15.*
